hinoko: move example programs
This commit moves example code from README to simplify it.
